server/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.2.0
Flask-Session==0.6.0
redis==5.0.1
gunicorn==21.2.0
